requirements.txt
Flask==3.0.1
pyserial==3.5